requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
faker>=24.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
